+++ CHANGELOG.md
@@ -5,7 +5,9 @@ SOUP version numbering generally follows the _major.minor.patch_ scheme.  Odd mi
 1.33 (development)
 ====
 
-...
+**New fixes in 1.32.0:**
+
+* [mso302336: Don't fail LQP review audits just for looking at the deletion popup](http://meta.stackoverflow.com/q/302336)
 
 1.32 (13 Aug 2015)
 ====